+++ packages/@expo/env/CHANGELOG.md
@@ -8,4 +8,6 @@
 
 ### 🐛 Bug fixes
 
+- Default to running in `development` mode when no `NODE_ENV` is specified.
+
 ### 💡 Others